More concise syntax
Smaller, slimmer, works better. The script should ALWAYS be in the same directory as "$EVBIN.original" because that's how the makefile sets things up. It is safe to assume that if $EVBIN.original exists, we can use the current path to the script as the base path, which is what `dirname $0` gets us.